costs from Long to Int

+++ library-api/src/main/scala/scalan/collection/Costs.scala
@@ -3,7 +3,7 @@ package scalan.collection
 trait Costed[Val] {
   def builder: CostedBuilder
   def value: Val
-  def cost: Long
+  def cost: Int
 }
 
 trait CostedBuilder {
